Replaces the switch/case for evaluating commands
Tried to be a little more clever. The switch/case was not ideal to select which code to run for each command. Now, all evaluation functions are stored in the map according to the command operation (e.g., SET). It's easier to add new commands, manage errors and arguments validation. But the code could be improved.